+++ priv/repo/migrations/20260408000000_add_metadata_updates.exs
@@ -1,7 +1,0 @@
-defmodule HephaestusOban.TestRepo.Migrations.AddMetadataUpdates do
-  use Ecto.Migration
-
-  def change do
-    HephaestusOban.Migration.add_metadata_updates()
-  end
-end